requirements.txt
# Core Dependencies
boto3>=1.34.0
psycopg2-binary>=2.9.9
strands-agents>=1.19.0

# Web UI
streamlit>=1.52.0

# Data Processing
pandas>=2.0.0
plotly>=5.18.0

# Optional: For development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
